+++ src/app/utils/to-raw-option-txt-list.ts
@@ -16,9 +16,11 @@ const weaponOptionSlots = [
   ItemOptionNumber.W_Right_1,
   ItemOptionNumber.W_Right_2,
   ItemOptionNumber.W_Right_3,
+  ItemOptionNumber.W_Right_4,
   ItemOptionNumber.W_Left_1,
   ItemOptionNumber.W_Left_2,
   ItemOptionNumber.W_Left_3,
+  ItemOptionNumber.W_Left_4,
 ];
 
 export const toRawOptionTxtList = (model: MainModel, itemMap) => {